requirements.txt
aioredis==2.0.0
asgiref==3.4.1
async-timeout==3.0.1
black==21.8b0
click==8.0.1
ecdsa==0.17.0
fastapi==0.68.1
h11==0.12.0
httptools==0.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
pyasn1==0.4.8
pydantic==1.8.2
python-dotenv==0.19.0
python-jose==3.3.0
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.28
rsa==4.7.2
six==1.16.0
starlette==0.14.2
tomli==1.2.1
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
motor
pymongo
